requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
psycopg2-binary>=2.9.9
bcrypt>=4.1.2
sqlalchemy>=2.0.0
scipy>=1.11.0
statsmodels>=0.14.0
